feat(userPanel): implement language management and localization support
