+++ userPanel/services/UserPanelViewService.ts
@@ -5,6 +5,7 @@ import ejs from "ejs";
 import path from "path";
 import { fileURLToPath } from "url";
 import { UserPanelAuthService } from "./UserPanelAuthService";
+import { UserPanelLanguageManager } from "./UserPanelLanguageManager";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
@@ -16,6 +17,7 @@ export class UserPanelViewService {
         private navigationService = ServiceContainer.getService(UserPanelNavigationService),
         private userPanelAuthService = ServiceContainer.getService(UserPanelAuthService),
         private translationManager = ServiceContainer.getService(TranslationManager),
+        private userPanelLanguageManager = ServiceContainer.getService(UserPanelLanguageManager),
     ) {}
 
     async render({
@@ -54,25 +56,8 @@ export class UserPanelViewService {
         const botName = this.client.user?.username || "Zumito";
         const tokenData = await this.userPanelAuthService.isLoginValid(req).then(result => result.data);
 
-        const availableLanguages = this.translationManager.getLanguages();
-        const defaultLanguage = this.translationManager.getDefaultLanguage();
-        let lang = req.cookies?.panel_lang;
-        if (!lang || !availableLanguages.includes(lang)) {
-            const header = req.headers['accept-language'] as string | undefined;
-            if (header) {
-                const parts = header.split(',').map(p => p.split(';')[0].trim());
-                lang = parts.map(p => p.slice(0, 2)).find(l => availableLanguages.includes(l));
-            }
-            if (!lang) lang = defaultLanguage;
-            res.cookie('panel_lang', lang, {
-                httpOnly: false,
-                secure: false,
-                sameSite: 'lax',
-                maxAge: 30 * 24 * 60 * 60 * 1000,
-                path: '/',
-            });
-        }
-        const t = this.translationManager.getShortHandMethod('', lang);
+        const { lang, t, availableLanguages, defaultLanguage } = this.userPanelLanguageManager.getLanguageVariables(req, res);
+
         return await ejs.renderFile(
             UserPanelViewService.layoutPath,
             {
